requirements.txt
torch
torchvision
torchaudio
openai
pillow
langchain==0.0.101
git+https://github.com/huggingface/transformers.git
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
git+https://github.com/facebookresearch/segment-anything.git
opencv-python==4.5.5.64
pycocotools
matplotlib
onnxruntime; sys_platform == "linux"
onnx; sys_platform == "linux"
https://gradio-builds.s3.amazonaws.com/3e68e5e882a6790ac5b457bd33f4edf9b695af90/gradio-3.24.1-py3-none-any.whl
accelerate
bitsandbytes
easyocr
tensorboardX